Add counter to waitForService, move queryIntentActivitiesAsUser into ServiceUtils
